test Unitaire class Abstraite Personnage
